requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
httpx==1.4.0
aiohttp==3.9.5
jinja2==3.1.4
python-dotenv==1.1.1
sse-starlette==1.7.3
